+++ config/application.rb
@@ -19,7 +19,7 @@ require "action_cable/engine"
 Bundler.require(*Rails.groups)
 
 module Backend
-  VERSION = "0.13.4"
+  VERSION = "0.14.0"
 
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.